requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
python-dotenv==1.0.1
supabase==1.2.0
gotrue==1.1.0
httpx>=0.24.1,<0.25.0
pydantic==2.6.1
pydantic-settings==2.1.0
PyJWT==2.8.0
openai==1.12.0
